+++ src/Components/Project.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { NavLink } from 'react-router-dom';
-import '../Project.css';
+import '../styling/Project.css';
 
 const Project = ({name}) => {
