+++ src/Components/Projects.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Project from './Project';
 import PropTypes from 'prop-types';
-import '../Projects.css';
+import '../styling/Projects.css';
 
 const Projects = ({category}) => {
   let allProjects = category.map(category => {
