+++ src/Components/Tries.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Project from './Components/Project';
+import Project from './Project';
 import PropTypes from 'prop-types';
 
 const Tries = ({category}) => {
